IMPALA-8369: Bump CDP_BUILD_NUMBER and re-enable test_max_nesting_depth
Switch to a newer version of CDP Hive where HIVE-20833 is reverted.
HIVE-20833 was backported without HIVE-20221, which broke the
the handling of column PARTITION_PARAMS.PARAM_KEY in HMS, leading
to several test failures in Impala.

The new HIVE version also includes the fix for HIVE-21796, so
test_max_nesting_depth could be also re-enabled with Hive 3.

Change-Id: I1d6f4e29997c9cf2238e1d614f8d1ed7d35ffe92