IMPALA-12430: Skip compression when sending row batches within the same process
LZ4 compression doesn't seem useful when the RowBatch is sent to a
fragment instance within the same process instead of a remote host.

See the Jira for discussion of tasks that could be skipped in
intra process RowBatch transfer. From these compression is both
the simplest to skip and the most expensive.

Note that it may also make sense to skip compression if the target
is not the same process but resides on the same host. This setup is
not typical in production environment AFAIK and it would complicate
testing of compression as impalad processes of often run on the
same host during tests. For these reasons it seems better to only
implement this if both the host and port are the same.

Change-Id: I7ea23fd1f0f10f72f3dbd8594f3def3ee190230a