IMPALA-12433: Share buffers among channels in KrpcDataStreamSender
Before this patch each KrpcDataStreamSender::Channel had 2
OutboundRowBatch with its own serialization and compression buffers.

This patch switches to use a single buffer per channel. This is
enough to store the in-flight data in KRPC, while other buffers
are only used during serialization and compression which is done for
just a single channel at a time, so can be shared among channels.

Memory estimates in the planner are not changed because the existing
calculation has several issues (see IMPALA-12594).

Change-Id: I64854a350a9dae8bf3af11c871882ea4750e60b3